+++ components/dasbootstrap/log/core.py
@@ -1,18 +1,9 @@
 """Logging system component."""
 
 import logging
-from cysystemd.journal import JournaldLogHandler
 
 # get an instance of the logger object this module will use
 logger = logging.getLogger(__name__)
-
-# instantiate the JournaldLogHandler to hook into systemd
-journald_handler = JournaldLogHandler()
-
-# set a formatter to include the level name
-journald_handler.setFormatter(logging.Formatter(
-    '[%(levelname)s] %(message)s'
-))
 
 # add the journald handler to the current logger
 logger.addHandler(journald_handler)
