+++ projects/automation_station/pyproject.toml
@@ -7,9 +7,8 @@ license = "MIT"
 
 packages = [
     {include = "dasbootstrap/automation_station", from = "../../bases"},
-    {include = "dasbootstrap/config", from = "../../components"},
+    {include = "dasbootstrap/resources", from = "../../components"},
     {include = "dasbootstrap/trapper_keeper", from = "../../components"},
-    {include = "dasbootstrap/files",from = "../../components"},
     {include = "dasbootstrap/keegen",from = "../../components"},
 ]
 
@@ -22,6 +21,7 @@ pykeepass = "^4.1.0"
 xdg-base-dirs = "^6.0.1"
 xkcdpass = "^1.19.9"
 zstandard = "^0.22.0"
+watchdog = "^4.0.1"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
