+++ projects/kiosk/pyproject.toml
@@ -1,0 +1,24 @@
+[tool.poetry]
+name = "kiosk"
+version = "0.1.0"
+description = ""
+authors = ["Jeremy Hettenhouser <jhettenh@gmail.com>"]
+license = "MIT"
+
+packages = [
+  { include = "dasbootstrap/chrome_kiosk", from = "../../bases" }
+]
+
+[tool.poetry.scripts]
+chrome_kiosk = "dasbootstrap.chrome_kiosk.core:chrome_kiosk"
+
+[tool.poetry.dependencies]
+python = "^3.12"
+playwright = "^1.45.0"
+pynput = "^1.7.7"
+undetected-playwright = "^0.3.0"
+fire = "^0.6.0"
+
+[build-system]
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
